// app/backend/src/types/TeamRank.ts
type TeamRank = {
  [key: string]: string | number;
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: number;
};

export default TeamRank;